refactor: move FlatOptions to rolldown_common (#6235)
Since `FlatOptions`​ are used in the most of stage of bundling move it to `rolldown_common`​ should be more proper.
