+++ crates/rolldown/src/ast_scanner/side_effect_detector/mod.rs
@@ -9,7 +9,7 @@ use oxc::ast::ast::{
 };
 use oxc::ast::{match_expression, match_member_expression};
 use oxc_allocator::Address;
-use rolldown_common::{AstScopes, SharedNormalizedBundlerOptions, SideEffectDetail};
+use rolldown_common::{AstScopes, FlatOptions, SharedNormalizedBundlerOptions, SideEffectDetail};
 use rolldown_utils::global_reference::{
   is_global_ident_ref, is_side_effect_free_member_expr_of_len_three,
   is_side_effect_free_member_expr_of_len_two,
@@ -29,124 +29,6 @@ bitflags! {
   pub struct PropertyAccessFlag: u8 {
     const Read = 1 << 0;
     const Write = 1 << 1;
-  }
-}
-
-bitflags! {
-  #[derive(Debug, Clone, Copy)]
-  /// A flat options struct to avoid passing `&SharedNormalizedBundlerOptions` everywhere.
-  /// which also make accessing frequently used options faster.
-  pub struct FlatOptions: u16 {
-    const IgnoreAnnotations = 1 << 0;
-    const JsxPreserve = 1 << 1;
-    const IsManualPureFunctionsEmpty = 1 << 2;
-    /// If the flag is set, it means the `treeshake.property_read_side_effects` is `Always`.
-    /// Otherwise, it is `False`.
-    const PropertyReadSideEffects = 1 << 3;
-    /// If the flag is set, it means the `treeshake.property_write_side_effects` is `Always`.
-    /// Otherwise, it is `False`.
-    const PropertyWriteSideEffects = 1 << 4;
-    /// If set, ESM import/export syntax should be preserved in the output.
-    /// Usage: `!self.options.format.keep_esm_import_export_syntax()`
-    const KeepEsmImportExportSyntax = 1 << 5;
-    /// If set, the format should call runtime require function.
-    /// Usage: `self.options.format.should_call_runtime_require()`
-    const ShouldCallRuntimeRequire = 1 << 6;
-    /// If set, polyfill require for ESM format when targeting Node platform.
-    /// Usage: `self.options.polyfill_require_for_esm_format_with_node_platform()`
-    const PolyfillRequireForEsmFormatWithNodePlatform = 1 << 7;
-    /// If set, new URL() calls with string literal and import.meta.url should be resolved to assets.
-    /// Usage: `self.options.experimental.is_resolve_new_url_to_asset_enabled()`
-    const ResolveNewUrlToAssetEnabled = 1 << 8;
-    /// If set, inline const optimization is enabled.
-    /// Usage: `self.options.optimization.is_inline_const_enabled()`
-    const InlineConstEnabled = 1 << 9;
-  }
-}
-
-impl FlatOptions {
-  pub fn from_shared_options(options: &SharedNormalizedBundlerOptions) -> Self {
-    let mut flags = Self::empty();
-    flags.set(Self::IgnoreAnnotations, !options.treeshake.annotations());
-    flags.set(Self::JsxPreserve, options.transform_options.is_jsx_preserve());
-    flags
-      .set(Self::IsManualPureFunctionsEmpty, options.treeshake.manual_pure_functions().is_none());
-    flags.set(
-      Self::PropertyReadSideEffects,
-      matches!(
-        options.treeshake.property_read_side_effects(),
-        rolldown_common::PropertyReadSideEffects::Always
-      ),
-    );
-    flags.set(
-      Self::PropertyWriteSideEffects,
-      matches!(
-        options.treeshake.property_write_side_effects(),
-        rolldown_common::PropertyWriteSideEffects::Always
-      ),
-    );
-    flags.set(Self::KeepEsmImportExportSyntax, options.format.keep_esm_import_export_syntax());
-    flags.set(Self::ShouldCallRuntimeRequire, options.format.should_call_runtime_require());
-    flags.set(
-      Self::PolyfillRequireForEsmFormatWithNodePlatform,
-      options.polyfill_require_for_esm_format_with_node_platform(),
-    );
-    flags.set(
-      Self::ResolveNewUrlToAssetEnabled,
-      options.experimental.is_resolve_new_url_to_asset_enabled(),
-    );
-    flags.set(Self::InlineConstEnabled, options.optimization.is_inline_const_enabled());
-    flags
-  }
-
-  #[inline]
-  pub fn ignore_annotations(self) -> bool {
-    self.contains(Self::IgnoreAnnotations)
-  }
-
-  #[inline]
-  pub fn jsx_preserve(self) -> bool {
-    self.contains(Self::JsxPreserve)
-  }
-
-  #[inline]
-  pub fn is_manual_pure_functions_empty(self) -> bool {
-    self.contains(Self::IsManualPureFunctionsEmpty)
-  }
-
-  #[inline]
-  pub fn property_read_side_effects(self) -> bool {
-    self.contains(Self::PropertyReadSideEffects)
-  }
-
-  #[inline]
-  pub fn property_write_side_effects(self) -> bool {
-    self.contains(Self::PropertyWriteSideEffects)
-  }
-
-  #[inline]
-  pub fn keep_esm_import_export_syntax(self) -> bool {
-    self.contains(Self::KeepEsmImportExportSyntax)
-  }
-
-  #[inline]
-  pub fn should_call_runtime_require(self) -> bool {
-    self.contains(Self::ShouldCallRuntimeRequire)
-  }
-
-  #[inline]
-  pub fn polyfill_require_for_esm_format_with_node_platform(self) -> bool {
-    self.contains(Self::PolyfillRequireForEsmFormatWithNodePlatform)
-  }
-
-  #[inline]
-  pub fn resolve_new_url_to_asset_enabled(self) -> bool {
-    self.contains(Self::ResolveNewUrlToAssetEnabled)
-  }
-
-  #[inline]
-  pub fn inline_const_enabled(self) -> bool {
-    self.contains(Self::InlineConstEnabled)
   }
 }
 
@@ -1081,7 +963,8 @@ mod test {
   use rolldown_common::{AstScopes, NormalizedBundlerOptions, SideEffectDetail};
   use rolldown_ecmascript::{EcmaAst, EcmaCompiler};
 
-  use super::{FlatOptions, SideEffectDetector};
+  use super::SideEffectDetector;
+  use rolldown_common::FlatOptions;
 
   fn get_statements_side_effect(code: &str) -> bool {
     let source_type = SourceType::tsx();
